# Refactor prometheus exporter to slightly improve performance