Depends on https://github.com/open-telemetry/opentelemetry-go/pull/3342

This removes the intermediary metricData struct for storing metrics, and instead directly constructs them and places them in the channel.

Before:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/exporters/prometheus
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkCollect1-16        	   34020	     36013 ns/op	   34377 B/op	      37 allocs/op
BenchmarkCollect10-16       	   17280	     71593 ns/op	   43729 B/op	     223 allocs/op
BenchmarkCollect100-16      	    3484	    315440 ns/op	  141521 B/op	    2049 allocs/op
BenchmarkCollect1000-16     	     403	   3019380 ns/op	 1178601 B/op	   20133 allocs/op
BenchmarkCollect10000-16    	      37	  31071324 ns/op	11247557 B/op	  200570 allocs/op
PASS
ok  	go.opentelemetry.io/otel/exporters/prometheus	7.537s
```

After:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/exporters/prometheus
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
BenchmarkCollect1-16        	   35424	     33992 ns/op	   34265 B/op	      34 allocs/op
BenchmarkCollect10-16       	   20886	     57156 ns/op	   42440 B/op	     198 allocs/op
BenchmarkCollect100-16      	    3814	    315079 ns/op	  129088 B/op	    1841 allocs/op
BenchmarkCollect1000-16     	     506	   2380297 ns/op	 1049404 B/op	   18121 allocs/op
BenchmarkCollect10000-16    	      45	  23002387 ns/op	 9850371 B/op	  180552 allocs/op
PASS
ok  	go.opentelemetry.io/otel/exporters/prometheus	7.234s
```