@dashpole 2022-10-19
This is tough to rebase.  I'll wait until after https://github.com/open-telemetry/opentelemetry-go/pull/3352 as well